+++ packages/etherplay-wallet-connector-ethereum/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @etherplay/wallet-connector-ethereum
 
+## 0.0.3
+
+### Patch Changes
+
+- wip: fuel connector
+- Updated dependencies
+  - @etherplay/wallet-connector@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes