+++ packages/etherplay-wallet-connector/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @etherplay/wallet-connector
 
+## 0.0.3
+
+### Patch Changes
+
+- wip: fuel connector
+
 ## 0.0.2
 
 ### Patch Changes